A function to convert MSnbase files to MSstatsQC format

A function to convert MSnbase files to MSstatsQC format

MSnbaseToMSstatsQC(msfile)

Arguments

msfile

data file to be converted

Value

A data frame that can be used with MSstatsQC

A csv file that is converted from raw files

Examples

# \donttest{
library(RforProteomics)
#> Loading required package: MSnbase
#> Loading required package: BiocGenerics
#> Warning: package 'BiocGenerics' was built under R version 4.5.2
#> Loading required package: generics
#> 
#> Attaching package: 'generics'
#> The following objects are masked from 'package:base':
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: 'BiocGenerics'
#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from 'package:base':
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> Loading required package: mzR
#> Loading required package: Rcpp
#> Warning: package 'Rcpp' was built under R version 4.5.2
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> 
#> Attaching package: 'S4Vectors'
#> The following object is masked from 'package:utils':
#> 
#>     findMatches
#> The following objects are masked from 'package:base':
#> 
#>     I, expand.grid, unname
#> Loading required package: ProtGenerics
#> Warning: package 'ProtGenerics' was built under R version 4.5.2
#> 
#> Attaching package: 'ProtGenerics'
#> The following object is masked from 'package:stats':
#> 
#>     smooth
#> 
#> This is MSnbase version 2.34.1 
#>   Visit https://lgatto.github.io/MSnbase/ to get started.
#>  Consider switching to the 'R for Mass Spectrometry'
#>  packages - see https://RforMassSpectrometry.org for details.
#> 
#> Attaching package: 'MSnbase'
#> The following object is masked from 'package:base':
#> 
#>     trimws
#> Warning: replacing previous import 'MSnbase::plot' by 'graphics::plot' when loading 'RforProteomics'
#> 
#> This is the 'RforProteomics' version 1.46.0.
#> 
#>   To get started, visit
#>     http://lgatto.github.com/RforProteomics/
#> 
#>   or, in R, open package vignettes by typing
#>     RforProteomics() # R/Bioc for proteomics overview
#>     RProtVis()       # R/Bioc for proteomics visualisation
#> 
#>   For a full list of available documents:
#>     vignette(package='RforProteomics')
#> 
#> 
#> Attaching package: 'RforProteomics'
#> The following object is masked from 'package:stats':
#> 
#>     spectrum
# }
# \donttest{
msfile <- getPXD000001mzXML()
#> Error in getPXD000001mzXML(): could not find function "getPXD000001mzXML"
# }
# \donttest{
MSnbaseToMSstatsQC(msfile)
#> Error: object 'msfile' not found
# }